+++ svalbard/xmlstream/src/main/java/org/n52/svalbard/write/WmlTVPEncoderv20XmlStreamWriter.java
@@ -263,7 +263,7 @@ public class WmlTVPEncoderv20XmlStreamWriter
         if (value != null && value.isSetValue()) {
             if (value instanceof QuantityValue) {
                 QuantityValue quantityValue = (QuantityValue) value;
-                return Double.toString(quantityValue.getValue());
+                return Double.toString(quantityValue.getValue().doubleValue());
             } else if (value instanceof ProfileValue) {
                 ProfileValue gwglcValue = (ProfileValue) value;
                 if (gwglcValue.isSetValue()) {
